+++ tests/ast_shape.py
@@ -6,8 +6,5 @@ print_ast = ast.Name("print", ast.Load())
 
 
 pprint(parse("""
-from a import b
-from . import c
-import b
-from . import *
+1 and 2
 """))